+++ docs/en/changelog.rst
@@ -23,11 +23,12 @@ Core
 
 * Can use IkParameterization to :meth:`.Robot.Manipulator.CheckEndEffectorCollision` even if it isn't a Transform6D type.
 
-* Added velocities enumerations to IkParameterizationType (like Transform6DVelocity), coordinate transforms can handle velocities.
-
-* :meth:`.IkSolver.Solve` can output reason of IK failures via :class:`.IkReturnAction` and custom data returned from IK filters through a :class:`.IkReturn` object.
+* Added velocities enumerations to IkParameterizationType (like IkParameterizationType.Transform6DVelocity), coordinate transforms can handle velocities.
 
 * Added IK solver error codes for notifying user where IK failed.
+
+* :meth:`.IkSolver.Solve` now returns results through the :class:`.IkReturn` object. It can output reason of IK failures (:class:`.IkReturnAction`) and custom data returned from IK filters.
+
 
 Database
 --------